fix serialization issue, fix app crash when home pressed twice, fix map screen imports, add .idea to git ignore

+++ app/src/main/java/abdulrahman/ali19/kist/data/pojo/model/weather/Minutely.kt
@@ -8,5 +8,5 @@ data class Minutely(
     @SerializedName("dt")
     val dt: Long,
     @SerializedName("precipitation")
-    val precipitation: Int
+    val precipitation: Double
 ) : Serializable
